Metadata of RaenonX-DL/dragalia-decompile@1c646df56fb1cf7da8d84f0e362a661c6def08f6

+++ Dumped/Assembly-CSharp/Gluon/HumanMarkUI.cs
@@ -129,7 +129,7 @@ namespace Gluon
 		public override void SetCircleGaugeUIGaugeTimer(CharaCircleGaugeUI.Type type, float time, float initialTime);
 		public override void SetCircleGaugeUICount(CharaCircleGaugeUI.Type type, int count);
 		[CompilerGenerated]
-		private void _Initialize_b__47_0(int cur, int next);
+		private void _Initialize_b__47_0(long cur, long next);
 		[CompilerGenerated]
 		private void _Initialize_b__47_1();
 		[CompilerGenerated]
